+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Bug Fixes
+
+- [boltdb] Use correct import path go.etcd.io/bbolt
+
 ## 0.5.0
 
 **2020-03-11**